adding endpoints

+++ src/main/java/cz/cvut/fit/tjv/service/CrudServiceImpl.java
@@ -35,6 +35,7 @@ public abstract class CrudServiceImpl<T extends EntityWithId<ID>, ID> implements
 
     @Override
     public void deleteById(ID id) {
+        if(!getRepository().findById(id).isPresent()) throw new RuntimeException("Id does not exist");
         getRepository().deleteById(id);
     }
